refactor(python): split binding helper owners
